pyproject.toml
[project]
name = "taller-agentes-universidad"
version = "0.1.0"
description = "Taller de Agentes - Ejemplos prácticos con LangChain y OpenAI"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai",
    "langchain-chroma",
    "langchain-text-splitters",
    "tiktoken",
    "pypdf",
    "python-dotenv",
    "streamlit",
    "langgraph>=0.2.39",
    "fpdf2>=2.8.7",
    "googlesearch-python>=1.3.0",
]

[dependency-groups]
dev = [
    "fpdf2>=2.8.7",
]
